+++ rollup-dev.config.js
@@ -2,6 +2,7 @@ import typescript from "@rollup/plugin-typescript";
 import { nodeResolve } from "@rollup/plugin-node-resolve";
 import commonjs from "@rollup/plugin-commonjs";
 import json from "@rollup/plugin-json";
+import copy from "rollup-plugin-copy";
 
 export default [
   {
@@ -17,12 +18,21 @@ export default [
         format: "cjs",
       },
     ],
-    external: ["axios", "p-queue", "axios-retry"],
+    external: [
+      "axios",
+      "p-queue",
+      "axios-retry",
+      "protobufjs",
+      "fast-xml-parser",
+    ],
     plugins: [
       typescript(),
       nodeResolve({ browser: false }),
       commonjs(),
       json(),
+      copy({
+        targets: [{ src: "src/assets/*", dest: "dist/assets" }],
+      }),
     ],
   },
 ];