+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Feature
 
-1. `video.download`函数不传递`ffmpegBinPath`的情况下使用`web`模式，最高支持 720p，且 mediaOptions 参数不会生效
+1. `video.download`函数不传递`ffmpegBinPath`的情况下使用`web`模式，最高支持 720p，且 mediaOptions 参数除 qn 外不会生效
 
 ## Bug Fix