+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 1. 修复下载进度条可能出现无穷大的 bug
 2. 修复`common.getAreas`失效的 bug
+3.
 
 ## Feature